Refactor styles
