+++ app/common/containers/pages/UpdateFactorSignInPage/index.js
@@ -1,30 +1,27 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import withStyles from 'withStyles';
 import { withRouter } from 'react-router';
 
 import { H1 } from 'components/Title';
 import Button, { ButtonsGroup } from 'components/Button';
+import { Main, Header, Article } from 'components/CenterLayout';
 
 import SignInForm from 'containers/forms/SignInForm';
 
 import { onSubmit } from './redux';
 
-import styles from './styles.scss';
-
 @withRouter
-@withStyles(styles)
 @connect(null, { onSubmit })
 export default class UpdateFactorSignInPage extends React.Component {
   render() {
     const { onSubmit = () => {}, location, router } = this.props;
 
     return (
-      <section className={styles.main} id="sign-in-page">
-        <header className={styles.header}>
+      <Main id="update-factor-page">
+        <Header>
           <H1>Зміна фактора авторизації</H1>
-        </header>
-        <article className={styles.form}>
+        </Header>
+        <Article>
           <SignInForm
             onSubmit={onSubmit}
             initialValues={{
@@ -36,8 +33,8 @@ export default class UpdateFactorSignInPage extends React.Component {
               Назад
             </Button>
           </ButtonsGroup>
-        </article>
-      </section>
+        </Article>
+      </Main>
     );
   }
 }
