--- app/common/containers/pages/UpdateFactorSignInPage/styles.scss
@@ -1,12 +1,0 @@
-.main {
-  text-align: center;
-
-  .header {
-    margin-bottom: 32px;
-  }
-
-  .form {
-    max-width: 420px;
-    margin: 0 auto;
-  }
-}
