+++ app/common/containers/pages/UpdateFactorSuccessPage/index.js
@@ -1,25 +1,22 @@
 import React from 'react';
-import withStyles from 'withStyles';
 
 import { H2 } from 'components/Title';
 import Button from 'components/Button';
+import { Main, Header, Article } from 'components/CenterLayout';
 
-import styles from './styles.scss';
-
-@withStyles(styles)
 export default class UpdateFactorSuccessPage extends React.Component {
   render() {
     return (
-      <section className={styles.main} id="success-page">
-        <header className={styles.header}>
+      <Main id="update-factor-success-page">
+        <Header>
           <H2>Фактор було успішно змінено!</H2>
-        </header>
-        <article className={styles.form}>
+        </Header>
+        <Article>
           <Button color="blue" to="/sign-in">
             Повернутися до входу
           </Button>
-        </article>
-      </section>
+        </Article>
+      </Main>
     );
   }
 }
